Introduce Projective Coordinate (#36)
This pull request improves a performance of scalar multiplication.

On each operation of scalar multiplication, it is necessary to convert coordinate from a affine coordinate to a projective coordinate and vice versa. On this conversion, inverse operation is necessary, but it's too high cost operation.

So we would rathar not use inverse operation. Projective coordinate was applied to all operations of scalar multiplication.